implemented verification

+++ src/main/java/com/iche/sco/service/drugServiceImpl/DrugServiceImpl.java
@@ -1,14 +1,19 @@
 package com.iche.sco.service.drugServiceImpl;
 
-import com.iche.sco.dto.request.DrugCreationRequest;
-import com.iche.sco.dto.request.DrugsUpdateRequest;
-import com.iche.sco.dto.response.ApiResponse;
-import com.iche.sco.dto.response.DrugsResponse;
-import com.iche.sco.dto.response.ResponseCode;
+import com.iche.sco.dto.drugs.request.DrugCreationRequest;
+import com.iche.sco.dto.drugs.request.DrugDeleteRequest;
+import com.iche.sco.dto.drugs.request.DrugsUpdateRequest;
+import com.iche.sco.dto.drugs.response.DrugsResponse;
+import com.iche.sco.enums.ResponseCode;
+import com.iche.sco.dto.globalResponse.APIResponse;
+import com.iche.sco.enums.Role;
 import com.iche.sco.exception.DrugNotFoundException;
 import com.iche.sco.exception.InternalServerError;
+import com.iche.sco.exception.InvalidCredentialsException;
 import com.iche.sco.model.Drugs;
+import com.iche.sco.model.Users;
 import com.iche.sco.respository.DrugRepository;
+import com.iche.sco.utils.UserVerification;
 import com.iche.sco.utils.ValidationUtils;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
@@ -23,20 +28,26 @@ import java.util.stream.Collectors;
 public class DrugServiceImpl implements DrugService {
     private final DrugRepository drugRepository;
     private final ValidationUtils<DrugCreationRequest> drugCreationRequestValidationUtils;
+    private final UserVerification userVerification;
 
 
     @Override
-    public ApiResponse<DrugsResponse> createDrug(DrugCreationRequest drugCreationRequest) {
+    public APIResponse<DrugsResponse> createDrug(DrugCreationRequest drugCreationRequest) {
         drugCreationRequestValidationUtils.validate(drugCreationRequest);
+        Users users = userVerification.validateLoginUser(drugCreationRequest.userEmail());
 
-        if(drugRepository.existsByName(drugCreationRequest.name())){
-            throw new InternalServerError("drugs with this name already exist");
+        if(!(users.getRole().equals(Role.VENDOR) || users.getRole().equals(Role.ADMIN))){
+            throw new InvalidCredentialsException("drugs with this name already exist");
         }
 
+        if(drugRepository.existsByName(drugCreationRequest.drugName())){
+            throw new InternalServerError("drugs with this name already exist");
+        }
         Drugs drugs = Drugs.builder()
                 .packs(drugCreationRequest.packs())
-                .name(drugCreationRequest.name())
+                .name(drugCreationRequest.drugName())
                 .price(drugCreationRequest.price())
+                .user(users)
                 .build();
         drugRepository.save(drugs);
 
@@ -46,12 +57,12 @@ public class DrugServiceImpl implements DrugService {
                 .packs(drugs.getPacks())
                 .price(drugs.getPrice())
                 .build();
-        return new ApiResponse<>(ResponseCode.CREATE_DRUG_REQUEST_SUCCESSFUL.getMessage(), ResponseCode.CREATE_DRUG_REQUEST_SUCCESSFUL.getStatusCode(),response);
+
+        return new APIResponse<>(ResponseCode.CREATE_DRUG_REQUEST_SUCCESSFUL.getMessage(), ResponseCode.CREATE_DRUG_REQUEST_SUCCESSFUL.getStatusCode(),response);
     }
 
-
     @Override
-    public ApiResponse<DrugsResponse> getSingleDrug(Long id) {
+    public APIResponse<DrugsResponse> getSingleDrug(Long id) {
 
         Drugs drugs = drugRepository.findById(id).orElseThrow(
                 ()-> new DrugNotFoundException("drug not found")
@@ -62,11 +73,11 @@ public class DrugServiceImpl implements DrugService {
                 .packs(drugs.getPacks())
                 .price(drugs.getPrice())
                 .build();
-        return new ApiResponse<>(ResponseCode.CREATE_DRUG_REQUEST_SUCCESSFUL.getMessage(), ResponseCode.CREATE_DRUG_REQUEST_SUCCESSFUL.getStatusCode(),response);
+        return new APIResponse<>(ResponseCode.CREATE_DRUG_REQUEST_SUCCESSFUL.getMessage(), ResponseCode.CREATE_DRUG_REQUEST_SUCCESSFUL.getStatusCode(),response);
     }
 
     @Override
-    public ApiResponse<List<DrugsResponse>> listOfDrugs() {
+    public APIResponse<List<DrugsResponse>> listOfDrugs() {
 
         List<Drugs> drugs = drugRepository.findAll();
         List<DrugsResponse> response = drugs.stream()
@@ -77,36 +88,44 @@ public class DrugServiceImpl implements DrugService {
                         .packs(drug.getPacks())
                         .build()
                 ).collect(Collectors.toList());
-        return new ApiResponse<>(ResponseCode.VIEW_DRUG_REQUEST_SUCCESSFUL.getMessage(), ResponseCode.VIEW_DRUG_REQUEST_SUCCESSFUL.getStatusCode(),response);
+        return new APIResponse<>(ResponseCode.VIEW_DRUG_REQUEST_SUCCESSFUL.getMessage(), ResponseCode.VIEW_DRUG_REQUEST_SUCCESSFUL.getStatusCode(),response);
     }
 
     @Override
-    public ApiResponse<String> deleteDrug(Long id) {
-        Drugs drugs = drugRepository.findById(id).orElseThrow(
+    public APIResponse<String> deleteDrug(DrugDeleteRequest drugDeleteRequest) {
+        Users users = userVerification.validateLoginUser(drugDeleteRequest.userEmail());
+
+        if(!(users.getRole().equals(Role.VENDOR) || users.getRole().equals(Role.ADMIN))){
+            throw new InvalidCredentialsException("drugs with this name already exist");
+        }
+        Drugs drugs = drugRepository.findById(drugDeleteRequest.id()).orElseThrow(
                 ()->  new InternalServerError("Drug not found"));
         drugRepository.delete(drugs);
-        return new ApiResponse<>(ResponseCode.VIEW_DRUG_REQUEST_SUCCESSFUL.getMessage(), ResponseCode.VIEW_DRUG_REQUEST_SUCCESSFUL.getStatusCode(),"drugs deleted");
+        return new APIResponse<>(ResponseCode.VIEW_DRUG_REQUEST_SUCCESSFUL.getMessage(), ResponseCode.VIEW_DRUG_REQUEST_SUCCESSFUL.getStatusCode(),"drugs deleted");
     }
 
     @Override
-    public ApiResponse<DrugsResponse> updateDrug(DrugsUpdateRequest drugsUpdateRequest) {
+    public APIResponse<DrugsResponse> updateDrug(DrugsUpdateRequest drugsUpdateRequest) {
+        Users users = userVerification.validateLoginUser(drugsUpdateRequest.userEmail());
 
-        if(!drugRepository.existsById(drugsUpdateRequest.id())){
-            throw new DrugNotFoundException("drug not found");
+        if(!(users.getRole().equals(Role.VENDOR) || users.getRole().equals(Role.ADMIN))){
+            throw new InvalidCredentialsException("drugs with this name already exist");
         }
-        Drugs drug = Drugs.builder()
-                .name(drugsUpdateRequest.name())
-                .price(drugsUpdateRequest.price())
-                .packs(drugsUpdateRequest.packs())
-                .build();
+        Drugs drug = drugRepository.findById(drugsUpdateRequest.id()).orElseThrow(
+                ()-> new DrugNotFoundException("drugs not found"));
+
+                drug.setName(drugsUpdateRequest.drugName());
+                drug.setPrice(drugsUpdateRequest.price());
+                drug.setPacks(drugsUpdateRequest.packs());
+
         drugRepository.save(drug);
 
         DrugsResponse response = DrugsResponse.builder()
                 .id(drugsUpdateRequest.id())
-                .name(drugsUpdateRequest.name())
+                .name(drugsUpdateRequest.drugName())
                 .price(drugsUpdateRequest.price())
                 .packs(drugsUpdateRequest.packs())
                 .build();
-        return new ApiResponse<>(ResponseCode.UPDATE_DRUG_REQUEST_SUCCESSFUL.getMessage(),ResponseCode.UPDATE_DRUG_REQUEST_SUCCESSFUL.getStatusCode(),response);
+        return new APIResponse<>(ResponseCode.UPDATE_DRUG_REQUEST_SUCCESSFUL.getMessage(),ResponseCode.UPDATE_DRUG_REQUEST_SUCCESSFUL.getStatusCode(),response);
     }
 }
